task: refactor Step 3 — move auth integration tests + fixtures to iam.user (Task 178)
Move the 10 auth integration tests (iam.auth → iam.user) and the 4 auth
fixtures (ActivationTokenFixture, AuthApi, AuthSessionFactory,
LoginRateLimitStatements) into iam.user.fixtures. Repoint every
iam.auth.fixtures import across the test tree (moved tests plus six
already-in-user tests that consume AuthSessionFactory/AuthApi). Drop the
auth.fixtures package-info (target user.fixtures follows the no-package-info
convention). spotless:apply removed the now same-package UserApi import in
ActivationTokenFixture and reordered the repointed imports. No iam.auth
references remain in the test tree.

Issue: #178
