+++ ProductSpecification/tasks/178-refactoring-collapse-iam-auth/progress.md
@@ -16,9 +16,9 @@ Type: refactoring
 - [x] refactor (cleanup: verify imports/static-analysis)
 
 ### Step 3: Move auth integration tests + fixtures iam.auth → iam.user
-- [~] refactor (move integration tests + fixtures package; update package + imports)
-- [ ] refactor (cleanup: verify imports/static-analysis)
+- [x] refactor (move integration tests + fixtures package; update package + imports)
+- [x] refactor (cleanup: verify imports/static-analysis)
 
 ### Step 4: Remove empty auth package and verify
-- [ ] refactor (delete auth package-info.java files + empty auth dirs)
+- [~] refactor (delete auth package-info.java files + empty auth dirs)
 - [ ] green-acceptance (full build: ArchitectureTest + ./mvnw verify green)
