--- src/test/java/by/iivanov/rpm/iam/auth/fixtures/package-info.java
@@ -1,4 +1,0 @@
-@NullMarked
-package by.iivanov.rpm.iam.auth.fixtures;
-
-import org.jspecify.annotations.NullMarked;
