+++ src/test/java/by/iivanov/rpm/iam/user/ActivateAccountCsrfIntegrationTest.java
@@ -1,7 +1,7 @@
-package by.iivanov.rpm.iam.auth;
+package by.iivanov.rpm.iam.user;
 
-import by.iivanov.rpm.iam.auth.fixtures.ActivationTokenFixture;
-import by.iivanov.rpm.iam.auth.fixtures.AuthApi;
+import by.iivanov.rpm.iam.user.fixtures.ActivationTokenFixture;
+import by.iivanov.rpm.iam.user.fixtures.AuthApi;
 import by.iivanov.rpm.testing.AbstractApplicationIntegrationTest;
 import org.intellij.lang.annotations.Language;
 import org.junit.jupiter.api.DisplayName;
