+++ src/test/java/by/iivanov/rpm/iam/user/ActivationTokenValidationIntegrationTest.java
@@ -1,7 +1,7 @@
-package by.iivanov.rpm.iam.auth;
+package by.iivanov.rpm.iam.user;
 
-import by.iivanov.rpm.iam.auth.fixtures.ActivationTokenFixture;
-import by.iivanov.rpm.iam.auth.fixtures.AuthApi;
+import by.iivanov.rpm.iam.user.fixtures.ActivationTokenFixture;
+import by.iivanov.rpm.iam.user.fixtures.AuthApi;
 import by.iivanov.rpm.testing.AbstractApplicationIntegrationTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
