+++ src/test/java/by/iivanov/rpm/iam/user/AuthLoginIntegrationTest.java
@@ -1,9 +1,9 @@
-package by.iivanov.rpm.iam.auth;
+package by.iivanov.rpm.iam.user;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 
-import by.iivanov.rpm.iam.auth.fixtures.AuthApi;
-import by.iivanov.rpm.iam.auth.fixtures.AuthSessionFactory;
+import by.iivanov.rpm.iam.user.fixtures.AuthApi;
+import by.iivanov.rpm.iam.user.fixtures.AuthSessionFactory;
 import by.iivanov.rpm.testing.AbstractApplicationIntegrationTest;
 import org.intellij.lang.annotations.Language;
 import org.junit.jupiter.api.DisplayName;
