+++ src/test/java/by/iivanov/rpm/iam/user/fixtures/AuthSessionFactory.java
@@ -1,4 +1,4 @@
-package by.iivanov.rpm.iam.auth.fixtures;
+package by.iivanov.rpm.iam.user.fixtures;
 
 import by.iivanov.rpm.testing.api.AssertionResponse;
 import by.iivanov.rpm.testing.session.SessionContext;
